Fix order of deletion problems with singletons
Linux seems to run destructors of singletons in shared libraries in an
unpredictable order, and atext() doesn't help. This registers
singletons in the Registry and only destroys them when the Registry is
destroyed. The MapNode, in its turn, holds a reference to the
Registry, which ought to delay its destruction until a safe time.